Refactor creation of comments.
Existing comments are now deleted from merge requests before new
comments will be added. Additionally, line comments are now placed
before the summary so that the summary will be always visible. When
GitLab fails to show a comment for a diff then the comment will be
written again as a commit note.